Reorganize actions in data-explorers store, add actions for sentiment-analyses
