+++ app/src/client/common/constants/actionTypes.ts
@@ -1,4 +1,5 @@
 export const SET_ARTICLE_DATA_BY_SLUG = 'SET_ARTICLE_DATA_BY_SLUG';
 export const SET_ARTICLE_DATA = 'SET_ARTICLE_DATA';
+export const SET_SENTIMENT_ANALYSES_BY_SLUG = 'SET_SENTIMENT_ANALYSES_BY_SLUG';
 export const SET_STOCK_DATA_ALL = 'SET_STOCK_DATA_ALL';
 export const SET_STOCK_DATA_SINGULAR = 'SET_STOCK_DATA_SINGULAR';
